services/ocr-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
paddleocr==2.7.3
paddlepaddle==2.6.2
opencv-python-headless==4.9.0.80
Pillow==10.2.0
numpy==1.24.3
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
